Size is not grandeur

+++ 0x00-hello_world/6-size.c
@@ -15,7 +15,7 @@ int main(void)
 	printf("size of a char: %lu byte(s)\n", (unsigned int)sizeof(a));
 	printf("size of an int: %lu byte(s)\n", (unsigned int)sizeof(b));
 	printf("size of a long int: %lu byte(s)\n", (unsigned int)sizeof(c));
-	printf("size of a long long int: %lu byte(s)\n", sizeof(d));
+	printf("size of a long long int: %lu byte(s)\n", (unsigned int)sizeof(d));
 	printf("size of a float: %lu byte(s)\n", (unsigned int)sizeof(f));
 	return (0);
 }
